General refactoring of MPC code and especially mpc_acados
