+++ safe_control_gym/controllers/mpc/mpc.py
@@ -8,11 +8,10 @@ from termcolor import colored
 from safe_control_gym.controllers.base_controller import BaseController
 from safe_control_gym.controllers.lqr.lqr_utils import discretize_linear_system
 from safe_control_gym.controllers.mpc.mpc_utils import (compute_discrete_lqr_gain_from_cont_linear_system,
-                                                        compute_state_rmse, get_cost_weight_matrix,
-                                                        reset_constraints, rk_discrete)
+                                                        get_cost_weight_matrix, reset_constraints,
+                                                        rk_discrete)
 from safe_control_gym.envs.benchmark_env import Task
 from safe_control_gym.envs.constraints import GENERAL_CONSTRAINTS, create_constraint_list
-from safe_control_gym.utils.utils import timing
 
 
 class MPC(BaseController):
@@ -27,9 +26,7 @@ class MPC(BaseController):
             warmstart: bool = True,
             soft_constraints: bool = False,
             soft_penalty: float = 10000,
-            terminate_run_on_done: bool = True,
             constraint_tol: float = 1e-6,
-            compute_initial_guess_method: str = 'ipopt',
             use_lqr_gain_and_terminal_cost: bool = False,
             init_solver: str = 'ipopt',
             solver: str = 'ipopt',
@@ -46,10 +43,8 @@ class MPC(BaseController):
             warmstart (bool): If to initialize from previous iteration.
             soft_constraints (bool): Formulate the constraints as soft constraints.
             soft_penalty (float): Penalty added in the cost function for soft constraints.
-            terminate_run_on_done (bool): Terminate the run when the environment returns done or not.
             constraint_tol (float): Tolerance to add to the constraint as sometimes solvers are not exact.
             additional_constraints (list): List of additional constraints.
-            compute_initial_guess_method (str): Method to compute the initial guess. Options: None, 'ipopt', 'lqr'.
             use_lqr_gain_and_terminal_cost (bool): Use the LQR ancillary gain and terminal cost in the MPC.
             init_solver (str): Solver to use for initial guess computation.
             solver (str): Solver to use for MPC optimization.
@@ -82,11 +77,9 @@ class MPC(BaseController):
         self.soft_constraints = soft_constraints
         self.soft_penalty = soft_penalty
         self.warmstart = warmstart
-        self.terminate_run_on_done = terminate_run_on_done
 
-        self.X_EQ = self.env.X_GOAL
-        self.U_EQ = self.env.U_GOAL
-        self.compute_initial_guess_method = compute_initial_guess_method
+        self.X_EQ = self.model.X_EQ
+        self.U_EQ = self.model.U_EQ
         self.use_lqr_gain_and_terminal_cost = use_lqr_gain_and_terminal_cost
         self.init_solver = init_solver
         self.solver = solver
@@ -120,6 +113,18 @@ class MPC(BaseController):
         '''Cleans up resources.'''
         self.env.close()
 
+    def reset_before_run(self, obs=None, info=None, env=None):
+        '''Reinitialize just the controller before a new run.
+
+        Args:
+            obs (ndarray): The initial observation for the new run.
+            info (dict): The first info of the new run.
+            env (BenchmarkEnv): The environment to be used for the new run.
+        '''
+        self.x_prev = None
+        self.u_prev = None
+        super().reset_before_run(obs, info, env)
+
     def reset(self):
         '''Prepares for training or evaluation.'''
         print(colored('Resetting MPC', 'green'))
@@ -136,11 +141,7 @@ class MPC(BaseController):
         self.set_dynamics_func()
         # CasADi optimizer.
         self.setup_optimizer(self.solver)
-        # Previously solved states & inputs, useful for warm start.
-        self.x_prev = None
-        self.u_prev = None
-
-        self.setup_results_dict()
+        self.reset_before_run()
 
     def set_dynamics_func(self):
         '''Updates symbolic dynamics with actual control frequency.'''
@@ -170,55 +171,6 @@ class MPC(BaseController):
                                          self.model.nx,
                                          self.model.nu,
                                          self.dt)
-
-    @timing
-    def compute_initial_guess(self, init_state, goal_states=None):
-        '''Compute an initial guess of the solution to the optimization problem.'''
-        if goal_states is None:
-            goal_states = self.get_references()
-        print(colored(f'computing initial guess using {self.compute_initial_guess_method}', 'green'))
-        if self.compute_initial_guess_method == 'ipopt':
-            self.setup_optimizer(solver=self.init_solver)
-            opti_dict = self.opti_dict
-            opti = opti_dict['opti']
-            x_var = opti_dict['x_var']  # optimization variables
-            u_var = opti_dict['u_var']  # optimization variables
-            x_init = opti_dict['x_init']  # initial state
-            x_ref = opti_dict['x_ref']  # reference state/trajectory
-
-            # Assign the initial state.
-            opti.set_value(x_init, init_state)  # initial state should have dim (nx,)
-            # Assign reference trajectory within horizon.
-            opti.set_value(x_ref, goal_states)
-            # Solve the optimization problem.
-            try:
-                sol = opti.solve()
-                x_val, u_val = sol.value(x_var), sol.value(u_var)
-            except RuntimeError:
-                print(colored('Warm-starting fails', 'red'))
-                x_val, u_val = opti.debug.value(x_var), opti.debug.value(u_var)
-            x_guess = x_val
-            u_guess = u_val
-        elif self.compute_initial_guess_method == 'lqr':
-            # initialize the guess solutions
-            x_guess = np.zeros((self.model.nx, self.T + 1))
-            u_guess = np.zeros((self.model.nu, self.T))
-            x_guess[:, 0] = init_state
-            # add the lqr gain and states to the guess
-            for i in range(self.T):
-                u = self.lqr_gain @ (x_guess[:, i] - goal_states[:, i]) + np.atleast_2d(self.model.U_EQ)[0, :].T
-                u_guess[:, i] = u
-                x_guess[:, i + 1, None] = self.dynamics_func(x0=x_guess[:, i], p=u)['xf'].toarray()
-        else:
-            raise Exception('Initial guess method not implemented.')
-
-        self.x_prev = x_guess
-        self.u_prev = u_guess
-
-        # set the solver back
-        self.setup_optimizer(solver=self.solver)
-
-        return x_guess, u_guess
 
     def setup_optimizer(self, solver='qrsqp'):
         '''Sets up nonlinear optimization problem.
@@ -307,7 +259,6 @@ class MPC(BaseController):
             'cost': cost
         }
 
-    @timing
     def select_action(self,
                       obs,
                       info=None
@@ -334,10 +285,6 @@ class MPC(BaseController):
         goal_states = self.get_references()
         opti.set_value(x_ref, goal_states)
 
-        if self.compute_initial_guess_method is not None and self.x_prev is None and self.u_prev is None:
-            x_guess, u_guess = self.compute_initial_guess(obs)
-            opti.set_initial(x_var, x_guess)
-            opti.set_initial(u_var, u_guess)  # Initial guess for optimization problem.
         if self.warmstart and self.x_prev is not None and self.u_prev is not None:
             # Shift previous solutions by 1 step
             x_guess = deepcopy(self.x_prev)
@@ -426,110 +373,3 @@ class MPC(BaseController):
                              'state_error': [],
                              't_wall': []
                              }
-
-    def run(self,
-            env=None,
-            render=False,
-            logging=False,
-            max_steps=None,
-            terminate_run_on_done=None
-            ):
-        '''Runs evaluation with current policy.
-
-        Args:
-            render (bool): If to do real-time rendering.
-            logging (bool): If to log on terminal.
-
-        Returns:
-            dict: evaluation statistics, rendered frames.
-        '''
-        if env is None:
-            env = self.env
-        if terminate_run_on_done is None:
-            terminate_run_on_done = self.terminate_run_on_done
-
-        self.x_prev = None
-        self.u_prev = None
-        obs, info = env.reset()
-        print('Init State:')
-        print(obs)
-        ep_returns, ep_lengths = [], []
-        frames = []
-        self.setup_results_dict()
-        self.results_dict['obs'].append(obs)
-        self.results_dict['state'].append(env.state)
-        i = 0
-        if env.TASK == Task.STABILIZATION:
-            if max_steps is None:
-                MAX_STEPS = int(env.CTRL_FREQ * env.EPISODE_LEN_SEC)
-            else:
-                MAX_STEPS = max_steps
-        elif env.TASK == Task.TRAJ_TRACKING:
-            if max_steps is None:
-                MAX_STEPS = self.traj.shape[1]
-            else:
-                MAX_STEPS = max_steps
-        else:
-            raise Exception('Undefined Task')
-        self.terminate_loop = False
-        done = False
-        common_metric = 0
-        while not (done and terminate_run_on_done) and i < MAX_STEPS and not (self.terminate_loop):
-            action = self.select_action(obs)
-            if self.terminate_loop:
-                print('Infeasible MPC Problem')
-                break
-            # Repeat input for more efficient control.
-            obs, reward, done, info = env.step(action)
-            self.results_dict['obs'].append(obs)
-            self.results_dict['reward'].append(reward)
-            self.results_dict['done'].append(done)
-            self.results_dict['info'].append(info)
-            self.results_dict['action'].append(action)
-            self.results_dict['state'].append(env.state)
-            self.results_dict['state_mse'].append(info['mse'])
-            self.results_dict['state_error'].append(env.state - env.X_GOAL[i, :])
-            common_metric += info['mse']
-            print(i, '-th step.')
-            print('action:', action)
-            print('obs', obs)
-            print('reward', reward)
-            print('done', done)
-            print(info)
-            print()
-            if render:
-                env.render()
-                frames.append(env.render('rgb_array'))
-            i += 1
-        # Collect evaluation results.
-        ep_lengths = np.asarray(ep_lengths)
-        ep_returns = np.asarray(ep_returns)
-        if logging:
-            msg = '****** Evaluation ******\n'
-            msg += 'eval_ep_length {:.2f} +/- {:.2f} | eval_ep_return {:.3f} +/- {:.3f}\n'.format(
-                ep_lengths.mean(), ep_lengths.std(), ep_returns.mean(),
-                ep_returns.std())
-        if len(frames) != 0:
-            self.results_dict['frames'] = frames
-        self.results_dict['obs'] = np.vstack(self.results_dict['obs'])
-        self.results_dict['state'] = np.vstack(self.results_dict['state'])
-        try:
-            self.results_dict['reward'] = np.vstack(self.results_dict['reward'])
-            self.results_dict['action'] = np.vstack(self.results_dict['action'])
-            self.results_dict['full_traj_common_cost'] = common_metric
-            self.results_dict['total_rmse_state_error'] = compute_state_rmse(self.results_dict['state'])
-            self.results_dict['total_rmse_obs_error'] = compute_state_rmse(self.results_dict['obs'])
-        except ValueError:
-            raise Exception('[ERROR] mpc.run().py: MPC could not find a solution for the first step given the initial conditions. '
-                            'Check to make sure initial conditions are feasible.')
-        return deepcopy(self.results_dict)
-
-    def reset_before_run(self, obs, info=None, env=None):
-        '''Reinitialize just the controller before a new run.
-
-        Args:
-            obs (ndarray): The initial observation for the new run.
-            info (dict): The first info of the new run.
-            env (BenchmarkEnv): The environment to be used for the new run.
-        '''
-        self.reset()
